+++ packages/dendron-cli/src/commands/importPod.ts
@@ -1,32 +1,120 @@
-import {
-  getAllImportPods,
-  podClassEntryToPodItemV4,
-  PodItemV4,
-} from "@dendronhq/pods-core";
+import { DendronError, DEngineClientV2, Stage } from "@dendronhq/common-all";
+import { getAllImportPods, PodUtils } from "@dendronhq/pods-core";
+import _ from "lodash";
+import path from "path";
 import yargs from "yargs";
-import { CommandCLIOpts, PodCLICommand } from "./pod";
+import { CLICommand } from "./base";
+import { fetchPodClassV4, PodSource } from "./pod";
+import {
+  setupEngine,
+  setupEngineArgs,
+  SetupEngineCLIOpts,
+  SetupEngineResp,
+} from "./utils";
 
-export class ImportPodCLICommand extends PodCLICommand {
-  static getPods() {
-    return getAllImportPods();
+export { CommandCLIOpts as ExportPodCLIOpts };
+
+type CommandCLIOpts = {
+  port?: number;
+  engine?: DEngineClientV2;
+  cwd?: string;
+  stage: Stage;
+  // custom
+  podId: string;
+  showConfig?: boolean;
+  genConfig?: boolean;
+  podSource: PodSource;
+  podPkg?: string;
+} & SetupEngineCLIOpts;
+
+type CommandOpts = CommandCLIOpts & {
+  podClass: any;
+  config: any;
+} & SetupEngineResp;
+
+type CommandOutput = void;
+
+export class ImportPodCLICommand extends CLICommand<
+  CommandOpts,
+  CommandOutput
+> {
+  constructor() {
+    super({
+      name: "importPod",
+      desc: "use a pod to import notes",
+    });
   }
 
-  static async buildArgs(args: yargs.Argv<CommandCLIOpts>) {
-    const podItems: PodItemV4[] = ImportPodCLICommand.getPods().map((p) =>
-      podClassEntryToPodItemV4(p)
-    );
-    return ImportPodCLICommand.buildArgsCore(args, podItems);
+  buildArgs(args: yargs.Argv<CommandCLIOpts>) {
+    super.buildArgs(args);
+    setupEngineArgs(args);
+    args.option("podId", {
+      describe: "id of pod to use",
+      requiresArg: true,
+    });
+    args.option("showConfig", {
+      describe: "show pod configuration",
+    });
+    args.option("genConfig", {
+      describe: "show pod configuration",
+    });
+    args.option("podPkg", {
+      describe: "if specifying a remote pod, name of pkg",
+    });
+    args.option("podSource", {
+      describe: "podSource",
+      choices: _.values(PodSource),
+      default: PodSource.BUILTIN,
+    });
   }
 
-  static async run(args: CommandCLIOpts) {
-    const ctx = "ImportPod:run";
-    const cmd = new ImportPodCLICommand();
-    cmd.L.info({ ctx, msg: "enter", args });
-    const pods = await ImportPodCLICommand.getPods();
-    const opts = await cmd.enrichArgs(args, pods, "import");
-    cmd.L.info({ ctx, msg: "enrichArgs:post", args });
-    await cmd.execute(opts);
-    cmd.L.info({ ctx, msg: "exit", args });
-    return cmd;
+  async enrichArgs(args: CommandCLIOpts): Promise<CommandOpts> {
+    const { podId, wsRoot, showConfig, podSource, podPkg, genConfig } = args;
+
+    const engineArgs = await setupEngine(args);
+    const pods = getAllImportPods();
+    const podType = "import";
+    const podClass = fetchPodClassV4(podId, {
+      podSource,
+      pods,
+      podType,
+      podPkg,
+      wsRoot,
+    });
+    if (showConfig) {
+      const config = new podClass().config;
+      console.log(config);
+      process.exit(0);
+    }
+    if (genConfig) {
+      const podsDir = PodUtils.getPodDir({ wsRoot });
+      const configPath = PodUtils.genConfigFile({ podsDir, podClass });
+      console.log(`config generated at ${configPath}`);
+      process.exit(0);
+    }
+    const podsDir = path.join(wsRoot, "pods");
+    const maybeConfig = PodUtils.getConfig({ podsDir, podClass });
+    if (!maybeConfig) {
+      const podConfigPath = PodUtils.getConfigPath({ podsDir, podClass });
+      throw new DendronError({
+        status: "no-config",
+        msg: `no config found. please create a config at ${podConfigPath}`,
+      });
+    }
+    return { ...args, ...engineArgs, podClass, config: maybeConfig };
+  }
+
+  async execute(opts: CommandOpts) {
+    const { podClass: PodClass, config, wsRoot, engine, server } = opts;
+    const vaults = engine.vaultsv3;
+    const pod = new PodClass();
+    console.log("running pod...");
+    await pod.execute({ wsRoot, config, engine, vaults });
+    server.close((err: any) => {
+      if (err) {
+        this.L.error({ msg: "error closing", payload: err });
+      }
+    });
+    console.log("done");
   }
 }