+++ packages/dendron-cli/src/commands/pod.ts
@@ -25,9 +25,11 @@ export function fetchPodClassV4(
   opts: {
     podSource: CommandCLIOpts["podSource"];
     pods?: PodClassEntryV4[];
+    podPkg?: string;
+    wsRoot?: string;
     podType: PodKind;
   }
-) {
+): PodClassEntryV4 {
   const { podSource, pods } = opts;
   if (podSource === "builtin") {
     if (!pods) {
@@ -36,10 +38,22 @@ export function fetchPodClassV4(
     const podClass = _.find(pods, {
       id: podId,
     });
+    if (_.isUndefined(podClass)) {
+      throw Error("no pod found");
+    }
     return podClass;
   } else {
-    const podEntry = require(podId);
-    const podClass = podEntry[opts.podType];
+    if (!opts.podPkg || !opts.wsRoot) {
+      throw Error("podPkg not defined");
+    }
+    const podEntries = require(`${path.join(
+      opts.wsRoot,
+      "node_modules",
+      opts.podPkg
+    )}`).pods as PodClassEntryV4[];
+    const podClass = _.find(podEntries, (entry) => {
+      return entry.id === podId && entry.kind === opts.podType;
+    });
     if (!podClass) {
       throw Error("no podClass found");
     }
@@ -47,11 +61,16 @@ export function fetchPodClassV4(
   }
 }
 
+export enum PodSource {
+  REMOTE = "remote",
+  BUILTIN = "builtin",
+}
+
 export type CommandCLIOpts = {
   podId: string;
   wsRoot: string;
   vault: DVault;
-  podSource?: "remote" | "builtin";
+  podSource?: PodSource;
 };
 
 export abstract class PodCLICommand extends BaseCommand<