+++ packages/engine-server/src/markdown/types.ts
@@ -9,6 +9,7 @@ import { Processor } from "unified";
 import { LinkFilter } from "../topics/markdown/plugins/types";
 import { DendronPubOpts } from "./remark/dendronPub";
 import { WikiLinksOpts } from "./remark/wikiLinks";
+export { VFile } from "vfile";
 
 export type DendronASTNode = Parent & {
   notes?: NotePropsV2[];
@@ -35,6 +36,7 @@ export type DendronASTData = {
   wikiLinkOpts?: WikiLinksOpts;
   config?: DendronConfig;
   overrides?: Partial<DendronPubOpts>;
+  shouldApplyPublishRules?: boolean;
 };
 
 // NODES